requirements.txt
pandas 
openai
python-dotenv
faiss-cpu
sentence-transformers
streamlit
